realworld-freesio-scala: Comments Services completed

+++ realworld-starter-kit/realworld-freesio-scala/src/main/scala/com/realworld/comments/persistence/CommentsQueries.scala
@@ -25,8 +25,8 @@ object CommentsQueries {
           LEFT JOIN users_followers uf on uf.follower_id = ${user_id} and uf.followee_id = c.author_id
        """.query[(CommentsEntity, ProfileEntity)]
 
-  def delete(comment_id: Long): Update0 =
-    sql"""DELETE from comments where id = ${comment_id}""".update
+  def deleteQuery(comment_id: Long, author_id: Long, article_id: Long): Update0 =
+    sql"""DELETE from comments where id = ${comment_id} and author_id = $author_id and article_id= $article_id""".update
 
   val createQuery: Update0 =
     sql"""
